Assistant: give the bot the donation guide's full body text (auto-synced)
The donation guide is a 4-tab, client-rendered page, so only the first
bracket is in the page HTML, and Next forbids server-rendering its
components (react-dom/server is blocked in routes). Instead:

- Move the bracket content into a shared, server-importable module
  (donation-guide/content.tsx). The page renders it unchanged — same DOM,
  same tab behavior (verified locally, all four brackets + switching).
- Extract its text by walking the component trees, no react-dom/server
  (lib/assistant/donation-guide.ts) — ~12KB of clean per-amount text.
- Inject it as a cached system block in the assistant + admin test-run
  routes, so the bot can answer donation questions with the guide's own
  recommendations. Regenerates each deploy, so it always reflects the guide.
